@@ -1,37 +1,39 @@
 {
-  "name": "my-app",
+  "name": "web",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@material-ui/core": "^4.11.0",
-    "@testing-library/jest-dom": "^4.2.4",
-    "@testing-library/react": "^9.3.2",
-    "@testing-library/user-event": "^7.1.2",
-    "@types/jest": "^24.9.1",
-    "@types/jwt-decode": "^2.2.1",
-    "@types/node": "^12.12.68",
-    "@types/react": "^16.9.53",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-redux": "^7.1.9",
-    "@types/react-router-dom": "^5.1.6",
-    "@types/redux": "^3.6.0",
-    "@types/redux-logger": "^3.0.8",
+    "@reduxjs/toolkit": "^1.9.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
+    "@types/axios": "^0.14.0",
+    "@types/jest": "^27.5.2",
+    "@types/node": "^16.18.10",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.9",
+    "@types/react-redux": "^7.1.24",
+    "@types/redux-logger": "^3.0.9",
+    "@types/redux-persist": "^4.3.1",
     "@types/redux-saga": "^0.10.5",
-    "axios": "^0.20.0",
-    "axios-observable": "^1.1.3",
-    "bootstrap": "^4.5.3",
-    "jwt-decode": "^3.0.0",
-    "node-sass": "^4.14.1",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
-    "react-redux": "^7.2.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "3.4.3",
-    "redux": "^4.0.5",
+    "autoprefixer": "^10.4.13",
+    "axios": "^1.2.1",
+    "history": "^5.3.0",
+    "postcss": "^8.4.20",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-redux": "^8.0.5",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.0",
+    "redux-devtools-extension": "^2.13.9",
     "redux-logger": "^3.0.6",
-    "redux-saga": "^1.1.3",
-    "redux-thunk": "^2.3.0",
-    "typescript": "^3.9.7"
+    "redux-observable": "^2.0.0",
+    "redux-persist": "^6.0.0",
+    "redux-saga": "^1.2.2",
+    "rxjs": "^7.8.0",
+    "tailwindcss": "^3.2.4",
+    "typescript": "^4.9.4",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -40,10 +42,10 @@
     "eject": "react-scripts eject"
   },
   "eslintConfig": {
-    "extends": "react-app",
-    "rules": {
-      "@typescript-eslint/no-unused-vars": "off"
-    }
+    "extends": [
+      "react-app",
+      "react-app/jest"
+    ]
   },
   "browserslist": {
     "production": [
